requirements.txt
pandas
scikit-learn
pyserial
joblib
